// admin-frontend/src/lib/mdc/types.ts
export interface RemoteNetwork {
  id: string;
  members: RemoteNetworkMember[];
  name?: string;
  siteId: string;
  workspaceId: string;
  virtualNetworkId: string;
  ipAssignmentPools?: RemoteNetworkIPAssignmentPool[];
  managedRoutes?: RemoteNetworkRoute[];
}
// ...
export interface RemoteNetworkMember {
  id: string;
  name?: string;
  description?: string;
  ipAddresses?: string[];
  online: boolean;
  authorized: boolean;
  created: string; // ISO DateTime
  lastOnline?: string;
  latency?: number;
  phyiscalIPAddress?: string; // Note: typo in API
  clientVersion?: string;
}

export interface RemoteNetworkIPAssignmentPool {
  ipRangeEnd: string;
  ipRangeStart: string;
}

export interface RemoteNetworkRoute {
  target: string;
  via?: string;
}